flask-on-docker/services/web/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
GeoAlchemy2==0.7.0
SQLAlchemy==1.3.15
psycopg2-binary==2.8.4
rasterio==1.1.3
